Refactor issue table to extract filters
