+++ assets/js/vue/components/Form/IssueFilter/StateFilter.vue
@@ -18,7 +18,8 @@
       >
     </custom-checkbox-field>
 
-    <custom-checkbox-field for-id="filter-state-is-closed" :label="$t('issue.state.closed')">
+    <custom-checkbox-field
+        for-id="filter-state-is-closed" :label="$t('issue.state.closed')">
       <input
           class="custom-control-input" type="checkbox" id="filter-state-is-closed"
           v-model="isClosed"
@@ -32,8 +33,7 @@
 
 <script>
 
-
-import CustomCheckboxField from '../Library/FormLayout/CustomCheckboxField'
+import CustomCheckboxField from '../../Library/FormLayout/CustomCheckboxField'
 export default {
   components: { CustomCheckboxField },
   emits: ['input'],
@@ -44,50 +44,31 @@ export default {
       isClosed: true,
     }
   },
-  props: {
-    defaultIsRegistered: {
-      default: true
-    },
-    defaultIsClosed: {
-      default: true
-    },
-    defaultIsResolved: {
-      default: true
-    },
-  },
   watch: {
-    isRegistered: function () {
-      this.updateState()
-    },
-    isResolved: function () {
-      this.updateState()
-    },
-    isClosed: function () {
+    state: function () {
       this.updateState()
     },
   },
-  methods: {
-    updateState: function () {
+  computed: {
+    state: function () {
       let state = 0;
-      if (this.isRegistered || this.minimalState >= 1) {
+      if (this.isRegistered) {
         state = state | 2;
       }
-      if (this.isResolved || this.minimalState >= 2) {
+      if (this.isResolved) {
         state = state | 4;
       }
-      if (this.isClosed || this.minimalState >= 4) {
+      if (this.isClosed) {
         state = state | 8;
       }
 
-      this.$emit('input', state)
+      return state > 0 ? state : null;
     }
   },
-  mounted() {
-    this.isRegistered = this.defaultIsRegistered
-    this.isResolved = this.defaultIsResolved
-    this.isClosed = this.defaultIsClosed
-
-    this.updateState()
+  methods: {
+    updateState: function () {
+      this.$emit('input', this.state)
+    }
   }
 }
 </script>
