+++ assets/js/vue/components/Form/IssueFilter/WasAddedWithClientFilter.vue
