+++ assets/js/vue/components/FoyerIssues.vue
@@ -11,7 +11,7 @@
 
     <issues-table
         :construction-site="constructionSite"
-        :force-state="1"
+        view="foyer"
         @selected="selectedIssues = $event" />
   </div>
 </template>
